requirements.txt
flask
flask-mail
flask-jwt
python-dotenv
sqlalchemy
pymysql
jwt
flask-cors